Start working on tests
