+++ brainlit/algorithms/__init__.py
@@ -4,3 +4,5 @@ from brainlit.algorithms import connect_fragments
 from brainlit.algorithms.connect_fragments import *
 from brainlit.algorithms import trace_analysis
 from brainlit.algorithms.trace_analysis import *
+from brainlit.algorithms import detect_somas
+from brainlit.algorithms.detect_somas import *
