+++ brainlit/algorithms/detect_somas/__init__.py
@@ -1,0 +1,1 @@
+from brainlit.algorithms.detect_somas.detect import *
