+++ data/upload/info
@@ -1,0 +1,59 @@
+{
+  "data_type": "uint16",
+  "num_channels": 1,
+  "scales": [
+    {
+      "chunk_sizes": [
+        [
+          132,
+          100,
+          52
+        ]
+      ],
+      "encoding": "raw",
+      "key": "9560.295454545456_9741.30875_31628.93269230769",
+      "resolution": [
+        9560.295454545456,
+        9741.30875,
+        31628.93269230769
+      ],
+      "size": [
+        1056,
+        800,
+        416
+      ],
+      "voxel_offset": [
+        0,
+        0,
+        0
+      ]
+    },
+    {
+      "chunk_sizes": [
+        [
+          132,
+          100,
+          52
+        ]
+      ],
+      "encoding": "raw",
+      "key": "19120.59090909091_19482.6175_63257.86538461538",
+      "resolution": [
+        19120.59090909091,
+        19482.6175,
+        63257.86538461538
+      ],
+      "size": [
+        528,
+        400,
+        208
+      ],
+      "voxel_offset": [
+        0,
+        0,
+        0
+      ]
+    }
+  ],
+  "type": "image"
+}
