+++ docs/reference/algs.rst
@@ -19,3 +19,10 @@ Fragment Generation
 
 Connect Fragments
 -----------------
+
+.. currentmodule:: brainlit.algorithms.detect_somas
+
+Soma Detection
+--------------
+
+.. autoapifunction:: find_somas
